# BackendApplication/src/main/resources/openapi/openapi.yaml
openapi: 3.0.1
info:
  title: WrocłArts API
  description: REST API documentation for the WrocłArts system managing artistic activities in Wrocław.
  termsOfService: http://www.example.com/terms/
  contact:
    email: contact@maxime-luce.fr
  license:
    name: Apache-2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: http://localhost:8080/WroclArts
    description: Local Spring Boot Development Server
...
paths:
...
  /admin/organizations/{orgId}/status:
    put:
      tags:
        - admin
      summary: verify, reject, suspend or ban an organization
      operationId: adminUpdateOrganizationStatus
      parameters:
        - name: orgId
          in: path
          description: Numeric ID of the organization to moderate.
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Status update details with optional reason
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminStatusUpdate'
          application/xml:
            schema:
              $ref: '#/components/schemas/AdminStatusUpdate'
        required: true
      responses:
        200:
          description: Status successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
            application/xml:
              schema:
                $ref: '#/components/schemas/Organization'
        400:
          description: Invalid input
          content: {}
        404:
          description: Organization not found
          content: {}
      security:
        - wroclarts_auth:
            - admin:platform
      x-codegen-request-body-name: adminStatusUpdate
  /admin/events/{eventId}:
    delete:
      tags:
        - admin
      summary: cancel or moderate an event violating ToS
      operationId: adminDeleteEvent
      parameters:
        - name: eventId
          in: path
          description: Numeric ID of the event to moderate.
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Event successfully removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
            application/xml:
              schema:
                $ref: '#/components/schemas/Event'
        400:
          description: Invalid eventId supplied
          content: {}
        404:
          description: Event not found
          content: {}
      security:
        - wroclarts_auth:
            - admin:platform
  /admin/products/{productId}:
    delete:
      tags:
        - admin
      summary: moderate a product violating ToS
      operationId: adminDeleteProduct
      parameters:
        - name: productId
          in: path
          description: Numeric ID of the product to moderate.
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Product successfully removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
            application/xml:
              schema:
                $ref: '#/components/schemas/Product'
        400:
          description: Invalid productId supplied
          content: {}
        404:
          description: Product not found
          content: {}
      security:
        - wroclarts_auth:
            - admin:platform
...
components:
  schemas:
...
    Event:
      required:
        - title
        - date
        - organizationsId
      type: object
      xml:
        name: Event
      properties:
        eventId:
          type: integer
          format: int64
        title:
          type: string
          example: Pottery Workshop
        date:
          type: string
          example: "2026-10-14"
        time:
          type: string
          example: "18:00"
        address:
          type: string
          example: Nadodrze
        category:
          type: string
          example: Crafts
        price:
          type: number
          format: double
          example: 120.0
        numberOfPlaces:
          type: integer
          example: 15
        type:
          type: string
          example: In-Person
        status:
          type: string
          example: Published
        isRegistered:
          type: boolean
          description: Is the current user registered
          example: false
...
    Organization:
      required:
        - email
        - krsNumber
        - nameOrga
      type: object
      xml:
        name: Organization
      properties:
        orgId:
          type: integer
          format: int64
        email:
          type: string
          example: contact@ceramicstudio.pl
        nameOrga:
          type: string
          example: Ceramic Studio Wrocław
        krsNumber:
          type: string
          example: "0000123456"
        subscriptionDate:
          type: string
          format: date
          example: "2026-04-10"
        status:
          type: string
          example: PendingVerification
...
    Product:
      required:
        - name
        - price
      type: object
      xml:
        name: Product
      properties:
        productId:
          type: integer
          format: int64
        name:
          type: string
          example: Ceramic Mug
        price:
          type: number
          format: double
          example: 45.0
        stockQuantity:
          type: integer
          format: int32
          example: 10
        status:
          type: string
          example: Active
...
    AdminStatusUpdate:
      required:
        - status
      type: object
      xml:
        name: AdminStatusUpdate
      properties:
        status:
          type: string
          example: Verified
        reason:
          type: string
          example: KRS successfully verified
...
  securitySchemes:
    wroclarts_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: http://localhost/oauth/dialog
          scopes:
            write:events: modify events records in your account
            read:events: read your events records
            write:organizations: modify organization records
            read:organizations: read organization records
            write:products: modify shop products
            read:products: read shop products
            write:orders: modify order status
            read:orders: read order status
            admin:platform: exclusive platform administrator actions (moderation, verification)